requirements.txt
packaging
torch>=1.8.0
torchvision>=0.9.0
numpy==1.26.4
matplotlib>=3.3.0
shap>=0.39.0
Pillow>=8.0.0
